+++ src/dcspy/utils.py
@@ -25,8 +25,8 @@ from PIL import ImageColor
 from psutil import process_iter
 from requests import get
 
-from dcspy.models import (CTRL_LIST_SEPARATOR, AnyButton, BiosValue, Color, ControlDepiction, ControlKeyData, DcsBiosPlaneData, DcspyConfigYaml, LcdMode,
-                          Release, RequestModel, get_key_instance)
+from dcspy.models import (CONFIG_YAML, CTRL_LIST_SEPARATOR, DEFAULT_YAML_FILE, AnyButton, BiosValue, Color, ControlDepiction, ControlKeyData, DcsBiosPlaneData,
+                          DcspyConfigYaml, LcdMode, Release, RequestModel, get_key_instance)
 
 try:
     import git
@@ -35,8 +35,6 @@ except ImportError:
 
 LOG = getLogger(__name__)
 __version__ = '3.6.1'
-CONFIG_YAML = 'config.yaml'
-DEFAULT_YAML_FILE = Path(__file__).parent / 'resources' / CONFIG_YAML
 
 with open(DEFAULT_YAML_FILE) as c_file:
     defaults_cfg: DcspyConfigYaml = yaml.load(c_file, Loader=yaml.SafeLoader)