+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Internal:
   * Improve checking for new releases (pydantic model) - @emcek
   * Only one branch stable version of DCS is preferred for new installs - @emcek
+  * Make Nuitka with Python 3.13 default when building executable - @emcek
 
 ## 3.6.1
 * Fix issue font size for G19 can be overwritten #394 - @emcek